[US16] Checkstyle violations.
