+++ social_network_api/src/main/java/com/danit/socialnetwork/MyDelegatedEncoder.java
@@ -1,4 +1,4 @@
-package com.danit.socialnetwork.security;
+package com.danit.socialnetwork;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
