+++ social_network_api/src/main/java/com/danit/socialnetwork/repository/UserRepository.java
@@ -3,7 +3,11 @@ package com.danit.socialnetwork.repository;
 import com.danit.socialnetwork.model.DbUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<DbUser, Integer> {
-  DbUser findByUsername(String username);
+  Optional<DbUser> findByUsername(String username);
+
+//  Optional<DbUser> findByUsername(String username);
 
 }
